requirements.txt
# AWS SDK
boto3==1.34.23
botocore==1.34.23
# Data Processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
# Database
PyMySQL==1.1.0
sqlalchemy==2.0.25
# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
colorama==0.4.6
tabulate==0.9.0
# Date/Time
python-dateutil==2.8.2
# HTTP/API
requests==2.31.0
urllib3==2.0.7
# Testing
pytest==7.4.3
pytest-cov==4.1.0
# Monitoring
psutil==5.9.6
# Data Generation
faker==20.1.0
